requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
finnhub-python==2.0.0
get-all-tickers==1.6
google-api-core==1.22.1
google-auth==1.20.1
google-cloud==0.34.0
google-cloud-language==1.3.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
inflection==0.5.0
joblib==0.16.0
keyboard==0.13.5
lxml==4.5.2
more-itertools==8.4.0
multitasking==0.0.9
nltk==3.5
numpy==1.19.1
pandas==1.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-firebase==1.2
pytz==2020.1
Quandl==3.5.2
regex==2020.7.14
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
tqdm==4.48.2
urllib3==1.25.10
websocket-client==0.57.0
yfinance==0.1.54
